Change java flex-templates classpath
