+++ flex-templates/java/regional_dlp_de_identification/pom.xml
@@ -20,7 +20,7 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.google.cloud.blueprints.datawarehouse</groupId>
+  <groupId>com.google.cloud.blueprints.datawarehouse.deid</groupId>
   <artifactId>regional-txt-dlp-bq-streaming</artifactId>
   <version>1.0.0</version>
