+++ flex-templates/java/regional_dlp_transform/src/main/java/com/google/cloud/blueprints/datawarehouse/transform/BigQueryReadTransform.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.blueprints.datawarehouse;
+package com.google.cloud.blueprints.datawarehouse.transform;
 
 import com.google.api.services.bigquery.model.TableRow;
 import com.google.auto.value.AutoValue;
